Fix pluralization and spacing in index.html
The word 'Log' was not being pluralized, probably an oversight or copy-paste issue

If only 1 project has no log, then the verb 'haben' will now be changed to 'hat'

When buttons get too big or numerous, they get moved to the next line. There will now be a small gap between the lines.