Create model for AG grades and update grading view
Since all grading activities are to be stored, create a new model for those. The model is accessible on the admin page.
The previous fields in Project are removed and replaced by properties which select the points and justification of the last activity

The database can't see the new properties for searches, so the selection of ungraded projects had to be changed.
For each relevant project, if the number of grading activities is more than 0, then it is graded, otherwise it is equal to 0 and ungraded.
Slight benefit is the now absence of the unintuitive '-1' default value for grades.

The grading form from the POV of the AGs hasn't changed.
Internally, the create view for grading activities replaces the update view for the project. Due to that, the project isn't directly accessible in the AGGradeForm.
The project is sent as a hidden field similarly to the secret, and upon reception is checked again. If the returned project isn't valid, an error message is shown.
The order_id itself isn't checked. If the need arises, then it could be sent like the secret too.

The AGGradeView is now a CreateView instead of an UpdateView in order to create a new grading activity upon each successful grading.
The project isn't directly accessible, but the order_id lets us retrieve it. Minor changes were necessary.
Also populate the hidden 'project' field with the project (becomes the internal id for the row in the database, so no leakage).

E-Mail notification is still missing