Computation of current grade changed
When no manual grade is entered by the Orga team, the last grade before the deadline passed is used.
If after the deadline a new grade is entered, the Orga team has to manually select the corresponding grading in the admin menu.

In order to show the AG what has been saved even after the deadline, two new properties return the most recent grading.

Also updated the information on the success page.