Correctly model grading activity requirements
The Orga team requires that by default only gradings made in time (i.e. before deadline) are valid.
If an AG wants to change the grading afterwards, the Orga needs to intervene to manually change the grading

Previously, there was a single model for grading activities which recorded all gradings.
This proved cumbersome as there isn't really an easy way to programmatically select only the gradings made before/after the deadline at the points of need.
The arrangement also doesn't reflect the different nature of gradings made before and after the deadline.

So the grading activities are now split into two models, those made before and those made after the deadline. Both models are essentially identical in data, and inherit from a single abstract base model, but different in behaviour.
Which grading activity goes into which model is decided in the AGGradeForm, where the correct model is decided upon saving the new object based on current time.

It's now particularly easy to choose the valid grading. The manual ag_grade option in the project only accepts gradings made after the deadline (conveniently ordered with the last one top), and is null otherwise.
If it is not null, that grading is chosen. Is it null, then simply use the latest grading in the table for those made before the deadline.
This is mainly important for the Views where the graded/ungraded projects have to be selected.

The most recent grading has to be looked up in both tables, though that's less of an issue.

Of course both models are listed on the Admin page