List all gradings
The gradings are divided into two lists, first showing the belated gradings, then the in-time ones.
If either list or both are empty, they aren't shown. The currently valid grade, if it exists, is marked in blue.

Grading justifications are shown in full - there shouldn't be that many gradings anyway. This also means that no detail page is necessary.

Once the list was done for the project overview, it was trivial to add it to the AG grade view and the success view.
The AG grade view also now shows the most recent (possibly not valid) grade and justification for easy editing.

Lastly, a small fix in the admin panel where the add project page sometimes failed when restarting the server