Remove OrgaLog fields and squash migrations
The fields handled and requires_attention are not necessary and as such are removed.
The logic that depended upon these fields can be safely removed too. This is mainly a concern in orgalog.html

The resulting migration files have been squashed together.
Like in other branches, one unnecessary migration was removed.

Orgalogs now need a specific permission to be viewed