Add models for timetracking and accounts for students
The time tracking uses three models, first the interval which describes the time period for which the hours are collected, second the entries which define for which kind of work the time was spent.
The third model are the categories set by the Orga that define which kinds of work can be done
As a rough structure, only TLs add intervals while students only fill in the hours spent per category.