Improve TimeIntervalEntryForm
The new purpose for this form is to correct general errors in the number of hours entered, e.g. typos.
For actually adding hours, there will be a better way in the next commits

Added form validation:
- Student in question is known, so no need to have in the form
- Checking that the given interval belongs to this student
- Total hours must be non-negative
- Some category needs to be selected

Subsequently, during saving, the object may or may not exist (though usually does), so search for it and replace the number of hours.