Add hours directly on detail page
Students now see a '+' next to their hours in each category.
Clicking on it unhides a number input field and a button to add numbers
The input field checks the value to be between 0 and 999 in steps of quarter hours
When clicking on the button, an ajax request is sent to an API endpoint with category and hours
Some server-side validation is performed before adding the hours to the previous total
Lastly, the new total is l10n-ed and returned to immediately update the website.

For corrections, particularly subtractions, a button to the edit-hours page has been added.

The TimeTable now includes the row and column argument from which the entry was generated.
The new entries are both 'None' for the header row and the first column.