Create TLs on first access
TLs will get a link to /lti/log.
When first accessing the tool via that link, the user is assumed to be a TL and a db entry is created.
The Orga then has to confirm them still, as TL is a powerful position.

Removed old function which always created a TL upon User creation