Control shown content based on role
Some websites are going to be shared between TLs, students and the Orga team. The shown content can be selected based on tl_content, student_content and orga_content
The selection is in base_tl.

The breadcrumbs for the Log Overview correctly shows the index page and "TL Log".
All other current descendents of base_tl show only "Übersicht" with a link to it in their breadcrumbs and will have to be changed later