Redirect students to tt project overview
The general overview with multiple projects is now reserved for TLs.

For students, who have only 1 project, this view would have been confusing anyway
Therefore, they are now redirected to their respective overviews - the url is still useful for cases where the project nr is not known in advance (like the navbar)

For one, this simplifies the view as only TLs need to be considered. This will also aid in the implementation of the next features