Improve structure of project details
The project details page takes infos from every subsystem in the BP-Tool. Therefore, it also depends on all of them

In a first step, split the page logically into two separate sections, the table and the tabs. They are rendered via inclusion_tags which allow for more extensive computations based on the project than a simple {% include %} tag
The latter section, the tabs, might be turned into tabs to better structure the page visually.

The ProjectGradesMixin was changed to be useful in the inclusion tag. Also improved computation of valid grade to return None and to prevent accidental collisions where both tables show a grade as valid