Dynamically create project info table
The dependency between the project info table and its contents is now inverted.

When wanting to add an info row to the project info table, decorate the tag with ProjectInfoTable.register.
The function then makes an appropriate inclusion_tag and registers it in the tags file.
The tag is also added to a list for the template to render. The priority that is given during registering
fixes the order of infos in the final list.

When registering a tag as a developer, the function takes in the project and returns the context for the given template.

The tag can be defined anywhere, allowing further separation of concerns
In addition, the project info table doesn't need to know in advance about all the info contained within it and thus has
less reason to change. Instead, it describes the high level design of the table.

This new design is fully incorporated and all necessary changes for it are done.