Dynamically render project info tabs
Each tab registers a description tag and a content tag which are then used to generate the tabs
Only the first tab is active. The order of tabs is determined by the priority

The registration is two-staged. First, a ProjectInfoTabTags is created to which the tags are added.
Then, the tags are added; after each time, the object checks whether all tags are there and if they are,
registers all of them.

This design decouples the tab view from its contents