Implement access control for tl logs
The TL log websites may only be accessed by TLs. This excludes specifically students.
Then use the new roles and their AC functions to better document the access control.

Reading existing logs is only allowed for the current TL of the group to which the log belongs.
Deleting or updating them is only allowed for the original author while they are the TL of the project. Later TLs do not have this permission.