Move log tab registration
The module needs to be imported into views.py in order to load the tags and register them.
The imports are unused and should not be used in any way

The module views.py was chosen as the tags are essentially more views that just happen to be
injected into a third site instead of their own.