Add guiding text for TLs
To support TLs with their logs, the Orga team can define guiding texts which include
questions of interest. TLs can consider these texts in their TL log.

Texts can be defined in the admin panel of the BP model. Though a separate model,
the templates are inlined to appear there. The templates are one-to-one with BPs.