Add more grades for pitches and docs
Students can hold pitches and have to document their work. These works are graded by the Orga.
The grades are now added to the models with corresponding access methods in Project

Grades and Projects have a one-to-one relationship (which is not enforced), so adding multiple may lead
to logical errors. In that sense, the code reserves the right to pick any of the available grade instances.
As such, grade instances are supposed to be modified in-place; logging may be added later though.

Points and justification are returned as strings for visualization, being empty if none is available
A special value method may be called to retrieve the raw number (with 0 as default if none is available)