Add new grades to general project overview
On the overview page for all projects, show the three grades seperated by "/".
Additionally, to keep them at fixed length, they are right-adjusted to a
constant length. Also, the default replaces each possible digit with "-"

For that to work, the whole html is wrapped in <pre></pre> tags.
As such, the text cannot be formatted more readable, but one look at the result
shows easily what it is doing.