Fix 404 when new BP is active
Since the access control only checked for the project in the active BP,
the old projects couldn't be accessed anymore once their BP wasn't active anymore.

With this fix, the access control always uses the BP of the registered user, so
they can always access their project as long as it actually exists; after that,
they will once again get a 404.