Refactor project overview
The project overview contains the key information about the projects. As such, it also depends on many subsystems.
This is particularly true of the project list.

Similarly to the project infos table and the tabs, the list is now refactored into a modular structure
that doesn't depend on its entries; rather, the entries register themselves and invert the dependency.

A notable difference to the previous approaches (and it's really closer to the tabs than the other table)
is the use of a formal specification (in the form of an enum) that specifies the necessary tags and the
exposed context (in this particular case either nothing or the project).

In addition, the function `register_pair` makes it easier to register the necessary tags without
using up too much (visual) space if no particular context is necessary.
To still get unique and readable names for the tags, the tag registration can take a common prefix
and generate the names from that. If none is provided, then the function names are used (previous behaviour).