Refactor dynamic tag registration generation
The generation is moved into a common function which also replaces
the code for the tabs registration.

The input (enum) specification is turned into a list of tag types
as enums by default collapse all members with the same value which
is a problem if we need to differentiate them.
Also, the list of registered entities cached after sorting for faster
access times on reload. Typically, the registration is done once at
startup.