Highlight close grades
Grades that are close (< 2 points) to the next higher grade are highlighted in the table.
Only the grade entries themselves are coloured. They are also only highlighted if all
grades are available, since partial grades may end up going either way.

To that end, two new methods were added to the project model:
- total_points (which returns the total of all grades
- grade_complete (which returns whether all grades are given)