Make assignment of peer groups more readable
At times, all teams in the BP have to be assigned
to small peer groups where they collaborate with other
projects on simple tasks. These assignments should be random.

The assignments must also conform to some constraints. These
include no two teams with the same TLs or AGs, because
1. some tasks involve the TLs and to some extent the AGs, and
2. the teams should have as diverse experiences as possible to
learn from one another.

The initial implementation of the assignments ensures the constraints
hold and is at least partially random, even if not uniformly so.
That said, it isn't very readable, and the constant db accesses
and overwrites set traps to later introduce difficult to debug
errors and db issues when the code is changed.

The improved implementation works comparatively simple:
1. Get all groups
2. Randomly shuffle the groups
3. Split them into virtual peer groups that only exist in the local scope
4. Check that each peer group conforms to the constrains
5. If not, go back to 2, else proceed
6. Create the real peer groups and store them in the database

The proof of constraint satisfaction and uniform random assignments
is trivial.

The only complication arises when after the given max_tries, no valid
solution has been found (the approach is probabilistic, as is the initial
implementation). This case, indicated by a returned trial number of max_tries,
still creates all the (imperfect) peer groups and then emits a message to
the admins which groups are not valid. It is then up to the admins to rerun
the assignment or to make local adjustments that may reduce the randomness to
get a solution that is 'good enough'.
This behaviour s essentially the same as in the initial implementation.

Large parts of the initial implementation have been reused to ensure similar
external behaviour. The most significant difference is the replacement of the
exchange mechanism with above statistical approach.

In a test run with 70 Groups and 24 TLs, a solution was usually found after
5-10 tries, with some rare occurrences of up to 25 tries. Current MAX_TRIES
is set at 1000.