Improve conversion between total_points and grades
While the students are graded with points in 0.25 intervals,
the actual grades are non-linear and discrete. This requires
a conversion step.

Since the grading steps are known multiples of 10, a fast
approach is to calculate the lower bound and use that value
for the conversion table. While this operation is O(1), it is
fragile and error-prone due to the use of modular arithmetic
and the requirement to finally yield exact integers.

The grading boundaries create a nicely increasing list, so a more general approach is to bisect it and find the lower and upper bounds. It's theoretically a bit slower with O(log(n)), but the small size of the grading table means that there is no significant impact.
This approach also makes it trivially simple to calculate the necessary points to the next grade - adding an infinite sentinel ensures that there is always an upper bound that can't be reached, so people with a close to perfect grade don't get marked as close to the next higher grade.

The database has to ensure that the points are in the valid range.