Fortify grade conversion
Using -infinity instead of 0 for the lower bound ensures that negative
points don't crash the server. This is similar to previous behaviour.

If the points are close to 100, there is no reason not to show it as close
to the next grade; this was only neessary to workaround the
previous conversion which wasn't as flexible.